+++ src/modules/all.less
@@ -16,5 +16,4 @@
 
 @import "./esl-alert/core.less";
 
-@import "./esl-select/core.less";
-@import "./esl-select-list/core.less";
+@import "./forms/all.less";
